functions was divided in two files
