+++ utils/ocr_utils.py
@@ -1,14 +1,12 @@
-import re
 import os
+import re
 import sys
-import glob
 import cv2
 import pdf2image
-import PyPDF2
 import pytesseract
 from shutil import Error
+from .pdf_utils import get_PDF_numPages, get_book_name, get_paragraph
 
-# Imprime el progreso de un proceso
 def drawProgressBar(percent, barLen=20):
     sys.stdout.write("\r")
     progress = ""
@@ -20,53 +18,20 @@ def drawProgressBar(percent, barLen=20):
     sys.stdout.write("[ %s ] %.2f%%" % (progress, percent * 100))
     sys.stdout.flush()
 
-# Consigue el nombre de las carpetas
-def get_dirs(path='.'):
-    all_files = os.listdir(path)
-    ls_dir = [name for name in all_files if os.path.isdir(os.path.join(path, name))]
-    return ls_dir
-
-# Consigue los nombres de los archivos PDF
-def get_PDFs(path='.'):
-    ls_pdfs = glob.glob(os.path.join(path, "*.pdf"))
-    ls_pdfs = list(map(lambda x: str(x).replace('\\', '/'), ls_pdfs))
-    return ls_pdfs
-
-# Consigue el número de páginas de un PDF
-def get_PDF_numPages(path='./test.pdf'):
-    pdfFile = open(path, 'rb')
-    pdfReader = PyPDF2.PdfFileReader(pdfFile)
-    num_pages = pdfReader.numPages
-    pdfFile.close()
-    return num_pages
-
-def get_book_name(book_path: str) -> str:
-    """Recibe el path de un libro y retorna su nombre sin la terminación .pdf"""
-    if ('/' in book_path) or ('\\' in book_path):
-        book_name = book_path.split('/')[-1]
-        book_name = book_name.split('\\')[-1]
-    else:
-        book_name = book_path
-    book_name = book_name.replace('.pdf', '')
-    return book_name
-
 def sanitize_name(name: str) -> str:
-    """Elimina caracteres especiales y espacios, dejando solo letras y números."""
     return re.sub(r'[^a-zA-Z0-9]', '', name)
 
-# Verifica si las imágenes ya existen
 def there_are_imgs(book_path: str) -> bool:
     book_img_name = get_book_name(book_path)
     book_img_name += '_page_1.jpg'
-    imgs_folder = 'book_imgs/'
+    imgs_folder = 'book_imgs'
     return os.path.isfile(os.path.join(imgs_folder, book_img_name))
 
-# Genera las imágenes del libro
 def create_book_images(book_path: str) -> bool:
     try:
         pages = pdf2image.convert_from_path(pdf_path=book_path, dpi=350)
         book_img_name = get_book_name(book_path)
-        imgs_folder = 'book_imgs/'
+        imgs_folder = 'book_imgs'
         num_pages = get_PDF_numPages(book_path)
         for i, page in enumerate(pages):
             current_page = i + 1
@@ -78,7 +43,6 @@ def create_book_images(book_path: str) -> bool:
         print(Error)
         return False
 
-# Separa la página en párrafos
 def mark_region(image_path: str):
     im = cv2.imread(image_path)
     gray = cv2.cvtColor(im, cv2.COLOR_BGR2GRAY)
@@ -104,7 +68,6 @@ def mark_region(image_path: str):
         last_y = y
     return line_items_coordinates
 
-# Realiza OCR sobre una imagen y retorna el texto
 def ocr_img(img_path: str) -> str:
     pytesseract.pytesseract.tesseract_cmd = r'C:\Program Files\Tesseract-OCR\tesseract.exe'
     paragraphs = mark_region(img_path)
@@ -120,57 +83,6 @@ def ocr_img(img_path: str) -> str:
             text += str(pytesseract.image_to_string(thresh1, config='--psm 6'))
     return text
 
-# Extrae el contenido de una página de un PDF
-def get_PDF_content(path='./test.pdf', page=0):
-    content = ''
-    pdfFile = open(path, 'rb')
-    pdfReader = PyPDF2.PdfFileReader(pdfFile)
-    num_pages = pdfReader.numPages
-    if 0 <= page < num_pages:
-        content += pdfReader.getPage(page).extractText()
-    content = content.strip().replace('\n\n', '\n')
-    pdfFile.close()
-    return content
-
-def has_selectable_text(path: str) -> bool:
-    try:
-        text = get_PDF_content(path, 0)
-        return bool(text and text.strip())
-    except Exception as e:
-        return False
-
-def write_text(content='', path='.', name='book_references.txt'):
-    file_path = os.path.join(path, name)
-    try:
-        with open(file_path, 'a', encoding='utf-8') as f:
-            f.write(content)
-            f.write('\n\n\n')
-        return True
-    except Exception as e:
-        print(e)
-        return False
-
-# Busca y devuelve párrafos donde se encuentra la frase clave
-def get_paragraph(phrase='', text=''):
-    paragraph = ''
-    phrase = str(phrase).lower()
-    text = str(text).lower()
-    resultados = text.split(phrase)
-    num_resultados = len(resultados)
-    if num_resultados > 1:
-        for i in range(num_resultados):
-            add_char = 100
-            if i != 0:
-                paragraph += phrase.upper()
-                paragraph += resultados[i][:add_char]
-                add_char = len(resultados[i]) - add_char
-                if add_char < 1:
-                    add_char = 0
-            if i != num_resultados - 1:
-                paragraph += resultados[i][-add_char:]
-    return paragraph
-
-# Escanea el libro completo con OCR y guarda el texto en archivos organizados por libro
 def scan_book_with_OCR(path: str, phrase: str):
     num_pages = get_PDF_numPages(path)
     book_name = get_book_name(path)
@@ -182,7 +94,6 @@ def scan_book_with_OCR(path: str, phrase: str):
     if not os.path.exists(base_txt_folder):
         os.makedirs(base_txt_folder)
     
-    # Verifica si el libro ya fue escaneado previamente
     if not os.path.exists(book_txt_folder):
         os.makedirs(book_txt_folder)
         
@@ -216,20 +127,6 @@ def scan_book_with_OCR(path: str, phrase: str):
                 book_content = f.read()
         else:
             book_content = ""
-        match = get_paragraph(phrase, text=book_content)
-        if match:
-            info += 'Página ' + str(current_page) + '\n'
-            info += '...' + match + '...\n\n'
-    return info
-
-# Escanea un libro completo sin OCR, usando texto seleccionable
-def scan_book(path, phrase):
-    num_pages = get_PDF_numPages(path)
-    info = ''
-    for page in range(num_pages):
-        current_page = page + 1
-        drawProgressBar(percent=current_page / num_pages)
-        book_content = get_PDF_content(path, page)
         match = get_paragraph(phrase, text=book_content)
         if match:
             info += 'Página ' + str(current_page) + '\n'
